Refine unified turn loop context and tool handling
