+++ docs/project_analysis.md
@@ -76,8 +76,15 @@
     - Centralized slash command prefix and range parsing inside the palette helper so the session reuses the shared utilities, backed by unit tests for cursor edge cases.
     - Extended the shared slash palette to drive paging and home/end navigation shortcuts, ensuring the session defers all keyboard handling to the helper with dedicated coverage.
     - Introduced shared fuzzy search utilities and richer slash command ranking so palette suggestions stay relevant for substring and keyword matches beyond simple prefixes.
-- [ ] Split unified turn loop into state management, UI interaction, and tool routing submodules with clear API boundaries.【F:src/agent/runloop/unified/turn.rs†L1-L200】
-- [ ] Relocate context pruning and summarization workflows into an `agent/runloop/context_manager.rs` service and integrate asynchronous pipelines for tool execution results.【F:src/agent/runloop/unified/turn.rs†L54-L75】
+- [x] Split unified turn loop into state management, UI interaction, and tool routing submodules with clear API boundaries.【F:src/agent/runloop/unified/state.rs†L1-L58】【F:src/agent/runloop/unified/ui_interaction.rs†L1-L223】【F:src/agent/runloop/unified/tool_routing.rs†L1-L142】【F:src/agent/runloop/unified/turn.rs†L1-L120】
+    - Extracted `SessionStats`, `CtrlCState`, and related turn lifecycle helpers into `state.rs`, isolating shared session bookkeeping from the orchestration loop.【F:src/agent/runloop/unified/state.rs†L1-L58】
+    - Moved presentation helpers (session status rendering, placeholder spinners) to `ui_interaction.rs`, reducing `turn.rs`'s direct coupling to Ratatui handle internals.【F:src/agent/runloop/unified/ui_interaction.rs†L1-L223】
+    - Relocated human-in-the-loop tool approval logic into `tool_routing.rs`, encapsulating prompt flows and policy updates behind a focused interface.【F:src/agent/runloop/unified/tool_routing.rs†L1-L142】
+    - Shifted streaming response rendering and reasoning state tracking to `ui_interaction.rs`, keeping turn orchestration focused on decision flow while the helper manages markdown output and cancellation-aware cleanup.【F:src/agent/runloop/unified/ui_interaction.rs†L200-L441】
+- [x] Relocate context pruning and summarization workflows into an `agent/runloop/context_manager.rs` service and integrate asynchronous pipelines for tool execution results.【F:src/agent/runloop/unified/context_manager.rs†L1-L111】【F:src/agent/runloop/unified/tool_pipeline.rs†L1-L70】
+    - Centralized conversation trimming, curated prompt assembly, and token budget resets inside a reusable `ContextManager`, replacing ad-hoc helpers in the turn loop.【F:src/agent/runloop/unified/context_manager.rs†L13-L108】【F:src/agent/runloop/unified/turn.rs†L210-L275】
+    - Introduced a dedicated `tool_pipeline` module that executes tools under timeout, surfaces stdout/metadata, and drives the refactored turn loop's success/failure handling without inline timeout logic.【F:src/agent/runloop/unified/tool_pipeline.rs†L1-L70】【F:src/agent/runloop/unified/turn.rs†L1385-L1539】
+    - Upgraded tool failure telemetry to emit structured `ToolExecutionError` payloads in the UI and conversation history, preserving metadata for ledger entries instead of plain error strings.【F:src/agent/runloop/unified/turn.rs†L1504-L1567】
 - [ ] Decompose `src/acp/zed.rs` into reusable configuration gating, permission handling, and workspace trust components with mockable interfaces for testing.【F:src/acp/zed.rs†L40-L121】
 - [ ] Add integration tests that replay ACP message flows using mocks to validate the refactored adapters.
 - [ ] Share streaming assembly and tool-call reconstruction utilities across providers via a `llm/providers/shared` module.【F:vtcode-core/src/llm/providers/openrouter.rs†L1-L120】
